+++ lib/services/bank_account_service.dart
@@ -1,6 +1,7 @@
 
 import 'dart:convert';
 import 'package:another_square/services/authentication_service.dart';
+import 'package:another_square/services/backend_extensions.dart';
 import 'package:another_square/square_models.dart';
 import 'package:http/http.dart' as http;
 
@@ -40,7 +41,7 @@ class BankAccountService {
     };
 
     Uri endpoint = Uri.https(
-        baseUrl, "/v2/bank-accounts", request.toJson());
+        baseUrl, "/v2/bank-accounts", request.toJson().toQueryParam());
 
     //print (endpoint.toString());