+++ lib/services/catalog_service.dart
@@ -1,6 +1,7 @@
 
 import 'dart:convert';
 import 'package:another_square/services/authentication_service.dart';
+import 'package:another_square/services/backend_extensions.dart';
 import 'package:another_square/square_models.dart';
 import 'package:http/http.dart' as http;
 
@@ -307,7 +308,7 @@ class CatalogService {
     };
 
     Uri endpoint = Uri.https(
-        baseUrl, "/v2/catalog/list", request.toJson());
+        baseUrl, "/v2/catalog/list", request.toJson().toQueryParam());
 
     //print (endpoint.toString());