+++ lib/services/giftcard_service.dart
@@ -1,6 +1,7 @@
 
 import 'dart:convert';
 import 'package:another_square/services/authentication_service.dart';
+import 'package:another_square/services/backend_extensions.dart';
 import 'package:another_square/square_models.dart';
 import 'package:http/http.dart' as http;
 
@@ -48,7 +49,7 @@ class GiftCardService {
     };
 
     Uri endpoint = Uri.https(
-        baseUrl, "/v2/gift-cards", request.toJson());
+        baseUrl, "/v2/gift-cards", request.toJson().toQueryParam());
 
     //print (endpoint.toString());
 
@@ -91,7 +92,7 @@ class GiftCardService {
     //print (endpoint.toString());
 
     var response = await
-    http.post(endpoint, body: request.toJson(), headers: headers);
+    http.post(endpoint, body: jsonEncode(request.toJson()), headers: headers);
 
     if (response.statusCode == 200) {
       print (jsonDecode(response.body));
@@ -326,7 +327,7 @@ class GiftCardService {
     };
 
     Uri endpoint = Uri.https(
-        baseUrl, "/v2/gift-cards/activities", request.toJson());
+        baseUrl, "/v2/gift-cards/activities", request.toJson().toQueryParam());
 
     //print (endpoint.toString());
 
@@ -369,7 +370,7 @@ class GiftCardService {
     //print (endpoint.toString());
 
     var response = await
-    http.post(endpoint, body:request.toJson(), headers: headers);
+    http.post(endpoint, body:jsonEncode(request.toJson()), headers: headers);
 
     if (response.statusCode == 200) {
       print (jsonDecode(response.body));