+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 ## 0.0.1
 
-* Initial release. Supports all apis except Payments.
+* Initial release. Supports all APIs except Payments.
 
 ## 0.0.2
 
-* Updated package description. 
+* Updated package description.
+
+## 0.0.3
+
+* Completed Payment APIs support.
+
+## 0.0.4
+
+* Fixed API posts.
+  
+## 0.0.5
+
+* Fixed query params.